requirements.txt
opencv-python
numpy
Pillow
pyobjc-core
pyobjc-framework-Quartz